Replace getenv with configs.
