+++ app/utils/authlogin.py
@@ -2,33 +2,35 @@ import os
 import jwt
 from jwt import PyJWTError
 from datetime import datetime, timedelta
-from fastapi import Request, HTTPException, status
+from fastapi import Request, status
 import app.schemas.common as resp_schemas
 from fastapi.responses import JSONResponse
+from app.providers.config import configs
 
 
-SECRET_KEY = os.getenv("SECRET_KEY")
-ALGORITHM = "HS256"
-COOKIE_NAME = "auth_token"
-AUTH_SERVER = os.getenv("AUTH_SERVER")
 
 class AuthMiddleware:
     def __init__(self):
-        if not SECRET_KEY:
+        self.SECRET_KEY = configs.SECRET_KEY
+        self.ALGORITHM = "HS256"
+        self.COOKIE_NAME = "auth_token"
+        self.AUTH_SERVER = configs.AUTH_SERVER
+
+        if not self.SECRET_KEY:
             raise ValueError("SECRET_KEY is missing from environment variables")
 
     def create_access_token(self, data: dict, expires_delta: timedelta = timedelta(minutes=30)):
         to_encode = data.copy()
         expire = datetime.now() + expires_delta
         to_encode.update({"exp": expire})
-        encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+        encoded_jwt = jwt.encode(to_encode, self.SECRET_KEY, algorithm=self.ALGORITHM)
         return encoded_jwt
 
     async def __call__(self, request: Request, call_next):
         if request.url.path.startswith("/api/v1/login/") or request.url.path.startswith("/api/v1/query/") or request.url.path in ["/docs", "/docs#", "/openapi.json", "/redoc"]:
             return await call_next(request)
 
-        token = request.cookies.get(COOKIE_NAME)
+        token = request.cookies.get(self.COOKIE_NAME)
 
         if not token:
             return JSONResponse(
@@ -43,7 +45,7 @@ class AuthMiddleware:
             )
 
         try:
-            payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+            payload = jwt.decode(token, self.SECRET_KEY, algorithms=[self.ALGORITHM])
             username = payload.get("sub")
             if username is None:
                 return resp_schemas.CommonResponse(
@@ -68,11 +70,11 @@ class AuthMiddleware:
         new_token = self.create_access_token(data={"sub": username})
         response = await call_next(request)
         response.set_cookie(
-            key=COOKIE_NAME,
+            key=self.COOKIE_NAME,
             value=new_token,
             httponly=True,
             max_age=3600,
             path="/",
-            domain=AUTH_SERVER
+            domain=self.AUTH_SERVER
         )
         return response
